+++ src/lib/frameworks/suite.ts
@@ -36,11 +36,13 @@ export interface ISuite extends Nugget {
     queue (selective: boolean): void
     error (selective: boolean): void
     idleQueued (selective: boolean): void
+    errorQueued (selective: boolean): void
     debrief (result: ISuiteResult, selective: boolean): Promise<void>
     persist (): ISuiteResult
     refresh (options: SuiteOptions): void
     setFresh (fresh: boolean): void
     isFresh (): boolean
+    countChildren (): number
     hasChildren(): boolean
     contextMenu (): Array<Electron.MenuItemConstructorOptions>
     getRunningOrder (): number | null