+++ src/lib/process/factory.ts
@@ -1,11 +1,17 @@
 import { find } from 'lodash'
-import { ProcessOptions, IProcess, DefaultProcess } from './process'
+import { ProcessId, ProcessOptions, IProcess, DefaultProcess } from './process'
 import { Runners } from './runners'
 import pool from './pool'
 
 export class ProcessFactory {
 
-    public static make (options: ProcessOptions): IProcess {
+    /**
+     * Make a new process according to the given options.
+     *
+     * @param options The options for the process we're making.
+     * @param poolId An optional id with which the newly made process will be added to the pool.
+     */
+    public static make (options: ProcessOptions, poolId?: ProcessId): IProcess {
 
         let spawned: IProcess | null = null
 
@@ -28,7 +34,7 @@ export class ProcessFactory {
 
         spawned = spawned || new DefaultProcess(options)
 
-        pool.add(spawned)
+        pool.add(spawned, poolId)
 
         return spawned
     }