+++ src/lib/process/pool.ts
@@ -1,24 +1,46 @@
-import { IProcess } from './process'
+import { ProcessId, IProcess } from './process'
 
+/**
+ * A pool of running processes.
+ */
 class ProcessPool {
-    public readonly processes: { [index: number]: IProcess } = {}
+    public readonly processes: { [type in ProcessId]: IProcess } = {}
 
-    public add (process: IProcess): void {
-        const id = process.getId()
+    /**
+     * Add a new process to the pool.
+     *
+     * @param process The process being pooled.
+     * @param id An optional id with which process will be added to the pool.
+     */
+    public add (process: IProcess, id?: ProcessId): void {
+
+        // If id was given, we'll use it, otherwise we'll
+        // try to get it from the process itself.
         if (!id) {
-            return
+            id = process.getId()
+
+            // If we can't acquire the id, don't pool it.
+            if (!id) {
+                return
+            }
         }
 
-        this.processes[id] = process
+        this.processes[id!] = process
 
+        // Once the process closes, remove it from the pool.
         process.on('close', () => {
-            if (typeof this.processes[id] !== 'undefined') {
-                delete this.processes[id]
+            if (typeof this.processes[id!] !== 'undefined') {
+                delete this.processes[id!]
             }
         })
     }
 
-    public findProcess(id: number): IProcess | undefined {
+    /**
+     * Find a pooled and running process using its id.
+     *
+     * @param id The id of the process trying to be found.
+     */
+    public findProcess(id: ProcessId): IProcess | undefined {
         return this.processes[id]
     }
 }